Refactor SagaTracker component and modularize utilities
- Simplify imports in `PageRoutes.tsx` by breaking multi-line imports.
- Refactor `SagaTracker.tsx` to improve readability and logic structure.
- Move `IndeterminateCheck` to `components/IndeterminateCheck.tsx`.
- Move `loadInitial` to `components/loadInitial.ts` for better modularity.
- Update state initialization in `SagaTracker` to use new `loadInitial`.
